tests no longer use __init__.py
